refactor: extract instruments app from blowcomotion monolith
