+++ instruments/management/commands/migrate_primary_instruments.py
