+++ instruments/templates/emails/instrument_rental_request_returned.txt
