+++ instruments/templates/instrument_library/list.html
