+++ instruments/templates/instrument_library/manage.html
