+++ instruments/templates/instrument_rental_patreon_updated.html
