+++ instruments/templates/instrument_rental_return.html
