+++ instruments/templates/instrument_rental_token_error.html
